Refactor QueryFilterModern component and related utilities
- Moved loadConceptSets and loadAtlasCohortDefinition functions to separate utility files for better organization.
- Updated loadConceptSets to accept parameters for dataset ID retrieval and state management.
- Refactored loadAtlasCohortDefinition to streamline concept set handling and improve error management.
- Removed redundant code and interfaces, consolidating logic for clarity and maintainability.
- Enhanced type safety by refining TypeScript interfaces and type guards.
